+++ src/params.cpp
@@ -102,6 +102,8 @@ void write_compression_params(std::ostream &out, const compression_params &cp) {
   }
   write_string(out, cp.read_info.assay);
   write_string(out, cp.read_info.assay_confidence);
+  write_bool(out, cp.encoding.barcode_sort);
+  out.write(byte_ptr(&cp.encoding.cb_len), sizeof(uint32_t));
 }
 
 void read_compression_params(std::istream &in, compression_params &cp) {
@@ -182,13 +184,24 @@ void read_compression_params(std::istream &in, compression_params &cp) {
       } else {
         cp.read_info.assay_confidence = "N/A";
       }
+      if (in.peek() != std::char_traits<char>::eof()) {
+        cp.encoding.barcode_sort = read_bool(in);
+        in.read(byte_ptr(&cp.encoding.cb_len), sizeof(uint32_t));
+      } else {
+        cp.encoding.barcode_sort = false;
+        cp.encoding.cb_len = 16;
+      }
     } else {
       cp.read_info.assay = "auto";
       cp.read_info.assay_confidence = "N/A";
+      cp.encoding.barcode_sort = false;
+      cp.encoding.cb_len = 16;
     }
   } else {
     cp.read_info.assay = "auto";
     cp.read_info.assay_confidence = "N/A";
+    cp.encoding.barcode_sort = false;
+    cp.encoding.cb_len = 16;
   }
 }