+++ src/spring.h
@@ -27,7 +27,9 @@ void compress(const std::string &temp_dir, const string_list &input_paths,
               const std::string &r3_path = std::string(),
               const std::string &i1_path = std::string(),
               const std::string &i2_path = std::string(),
-              const std::string &assay_type = "auto");
+              const std::string &assay_type = "auto",
+              const std::string &cb_source_path = std::string(),
+              uint32_t cb_len = 16);
 
 void decompress(const std::string &temp_dir, const string_list &input_paths,
                 const string_list &output_paths, const int num_thr,